v2.0.0.3 - Fixed bug with convert dose relating to not updating the XML tree, not showing the header units when initializing, and past template references being retained in memory and causing subscripting/event errors.

+++ Optimate/Models/InstructionModelTest.cs
@@ -11,7 +11,7 @@ namespace OptiMate.Models
 
         public string DefaultInstructionTargetId => throw new NotImplementedException();
 
-        public ushort? DoseLevel { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public double? DoseLevel { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public int Index { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public ushort? InfMargin { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         public string InstructionTargetId { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
